refactor : post service
adapter, repository 추상클래스, Post Service 구현
