--- src/service/notion/index.ts
@@ -1,5 +1,0 @@
-export { getPostData as notionPostData } from "./posts/getPostData";
-export { getPostInfo as notionPostInfo } from "./posts/getPostInfo";
-export { getPosts as notionPosts } from "./posts/getPosts";
-export { getPostsFromTag as notionPostsFromTag } from "./posts/getPostsFromTag";
-export { getTags as notionTags } from "./posts/getTags";
