--- src/service/notion/utils/index.ts
@@ -1,1 +1,0 @@
-export { extractData } from "./extractData";
