+++ src/service/notion/utils/isFullPageResponse.ts
@@ -1,0 +1,15 @@
+import { isFullPageOrDatabase } from "@notionhq/client";
+import {
+  DatabaseObjectResponse,
+  PageObjectResponse,
+  PartialDatabaseObjectResponse,
+  PartialPageObjectResponse,
+} from "@notionhq/client/build/src/api-endpoints";
+
+export const isFullPageResponse = (
+  response:
+    | PageObjectResponse
+    | PartialPageObjectResponse
+    | PartialDatabaseObjectResponse
+    | DatabaseObjectResponse
+): response is PageObjectResponse => isFullPageOrDatabase(response) && response.object === "page";
